+++ CHANGELOG.md
@@ -13,10 +13,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - added Opensearch serverless module
 - updated replication module to avoid docker pull rate limits and resource creation race conditions
+- added RDS database module
 
 ### **Changed**
 
 ### **Removed**
+
+- removed SageMaker Studio module that moved to [MLOps Modules](https://github.com/awslabs/mlops-modules)
 
 =======